Added DECLARE, ADD, SUBTRACT instruction wired to symbol table.
Finished print(msg) interpolation

Finished randomized generation

Finished process memory log data structure

PROBLEM: Will not compile until LogUtils.cpp line 61 areguments changes to fit with the newly addedLogEntry class

PROBLEM: SLEEP and FOR instructions do not work properly in createProcess so please check how that was implemented

+++ SymbolTable.h
@@ -6,11 +6,11 @@
 class SymbolTable
 {
 public: 
-	void setVariable(const std::string& name, int value) { 
+	void setVariable(const std::string& name, uint16_t value) { 
 		table[name] = value; 
 	} 
 	
-	int getVariable(const std::string& name) {
+	uint16_t getVariable(const std::string& name) {
 		if (table.find(name) != table.end()) { 
 			return table[name]; 
 		}
